refactor(planner): extract custom hooks for history, autosave, map tiles, and canvas events
Moves ~450 lines of state management and event handling out of
traffic-control-planner.tsx into four focused custom hooks:

- useHistory: undo/redo stack with pushHistory/resetHistory/setObjects
- useAutosave: localStorage persistence effect
- useMapTiles: Mercator tile computation + image loading cache
- useCanvasEvents: all mouse/wheel handlers (toWorld, trySnap, hitTest)

Closes #192 (partial — PR 3 of 3)
